add router Redirect in App.js
