+++ src/components/HomePage/HomePage.css
@@ -1,5 +1,5 @@
 /* body {
-	background: #7F7FD5;  
+	background: #7F7FD5;
 	background: -webkit-linear-gradient(to bottom, #91EAE4, #86A8E7, #7F7FD5);
-	background: linear-gradient(to bottom, #91EAE4, #86A8E7, #7F7FD5); 
+	background: linear-gradient(to bottom, #91EAE4, #86A8E7, #7F7FD5);
 	} */
