+++ src/components/HomePage/sortTable.js
@@ -1,0 +1,39 @@
+import React from "react";
+
+ const useSortTableData = (items, config = null) => {
+
+	const [sortConfig, setSortConfig] = React.useState(config);
+
+	const sortedItems = React.useMemo(() => {
+		let sortableItems = [...items];
+
+		if (sortConfig !== null) {
+			sortableItems.sort((a, b) => {
+				if (a[sortConfig.key] < b[sortConfig.key]) {
+					return sortConfig.direction === 'ascending' ? -1 : 1;
+				}
+				if (a[sortConfig.key] > b[sortConfig.key]) {
+					return sortConfig.direction === 'ascending' ? 1 : -1;
+				}
+				return 0;
+			});
+		}
+		return sortableItems;
+	}, [items, sortConfig]);
+
+	const requestSort = (key) => {
+		let direction = 'ascending';
+		if (
+				sortConfig &&
+				sortConfig.key === key &&
+				sortConfig.direction === 'ascending'
+		) {
+			direction = 'descending';
+		}
+		setSortConfig({ key, direction });
+	};
+
+	return { items: sortedItems, requestSort, sortConfig };
+};
+
+export default useSortTableData;
